Add animation to the player spaceship

+++ common/inc/Packet/StartGame.hpp
@@ -20,7 +20,7 @@ namespace Packet {
   private:
     std::string   _ip;
     uint16_t      _port;
-    uint8_t	  _playerId;
+    uint16_t	  _playerId;
   };
 };
